Comma change
The comma there seems to unnecessarily break the sentence.

Also, in that section, this sentence is just confusing in this context:

"If your app provides an route:favorites, the framework will invoke it before rendering the template. Yes, this is a bit repetitive."

Huh? Did I miss something on how something like a "route:favorites" is declared. If I understood that, I might understand what is repetitive.