Fix incorrect reference to `rentalUnit` variable.
This paragraph references a `rentalUnit` variable that's passed into the component, but the correct name of the variable is `rental`. This updates the paragraph to match the code.

A clearer option could be to leave the paragraph as-is, and *actually* name the variable `rentalUnit` in the code here and in `model-hook.md`.