Grammatical fixes (#1526)
* Write out "two" in a sentence

When used in a sentence, "2" is usually written out as "two" instead.

* Make wording of list items more parallel

This sentence gives a list of two actions, but previously the two were worded differently in a confusing way: the first said "one to…", and the other said "and an action to…". The relationship between the two is clearer if the phrasing is parallel: "one action to…", "and the other action to…".

* Remove extraneous "Both" in sentence

The sentence starts with "Both", but there aren't two things being referred to: just the one `filter` action. The "Both" was probably a holdover from a previous version where there _were_ two actions passed in.

* Add missing period at end of bulleted item

The other bulleted items in this list are all sentences that end with a period.