Explain how computed properties work
The text has been corrected to say that the function backing the
computed property is run once,  and the resulting value is cached.
Subsequent calls read from the cache, until any of the dependent
properties change, where the function has to be run again to update the
cache.

Replace `var` with `let`

Reword how cache affects computed functions

Thanks @toddjordan