Reword notes about asynchrony
I'm not convinced that my wording is ideal, but I wanted to point out that it wasn't very clear to me - what was being said.

Side note: the line numbers don't seem to be helpful, and in many cases - (take the 2 tests on this page) are actually more confusing. It's hard to tell where that second test goes since it should be line 45 based on the previous snippet. Also - it's valuable space. Anyway... didn't see a good way to mention that.