Show how to generate routes with Ember CLI (#1575)
This change shows how to generate routes using Ember CLI. It shows how
to generate a basic route and a nested route too. I thought of using the
`--path` option to specify the paths but looks like CLI has a bug where,
for a nested route, if the parent path is not already defined, it uses
the same path the parent and all child routes.